+++ app/Http/Controllers/DetaildevisController.php
@@ -41,9 +41,8 @@ class DetailDevisController extends Controller
     public function store(Request $request)
     {
 
-        if(Auth::check()){
-
-            $agence_id=Auth::user()->agence_id;
+        if (Auth::check()) {
+            $agence_id = Auth::user()->agence_id;
 
             // dd(
             // $request->devis_id,
@@ -52,56 +51,49 @@ class DetailDevisController extends Controller
             // $agence_id,
             // $request->qte,
             // $request->prix,);
-// dd(StockProduitActivite::where('produit_id',$request->produit)->where('activite_id',$request->activite_id)->where('agence_id',$agence_id)->first()->id);
+            // dd(StockProduitActivite::where('produit_id',$request->produit)->where('activite_id',$request->activite_id)->where('agence_id',$agence_id)->first()->id);
 
 
-        if(isset(StockProduitActivite::where('produit_id',$request->produit)->where('activite_id',$request->activite_id)->where('agence_id',$agence_id)->first()->id))
-        {
-            $stock=StockProduitActivite::where('produit_id',$request->produit)->where('activite_id',$request->activite_id)->where('agence_id',$agence_id)->first();
-                if($stock->quantite_en_stock>=$request->qte)
-                {
-                    if(isset(DetailDevis::where('devis_id',$request->devis_id)->where('produit_id',$request->produit)->first(['id'])->id))
-                    {
-                        return back()->with('danger',"Produit deja enregistrer");
-                    }else{
-                         $data=[
-                                'devis_id'              =>$request->devis_id,
-                                'produit_id'               =>$request->produit,
-                                'quantite_demandee'       =>$request->qte,
-                                'prix_unitaire_demande'  =>$request->prix,
-                            ];
-                         DetailDevis::create($data);
+            if (isset(StockProduitActivite::where('produit_id', $request->produit)->where('activite_id', $request->activite_id)->where('agence_id', $agence_id)->first()->id)) {
+                $stock = StockProduitActivite::where('produit_id', $request->produit)->where('activite_id', $request->activite_id)->where('agence_id', $agence_id)->first();
+                if ($stock->quantite_en_stock >= $request->qte) {
+                    if (isset(DetailDevis::where('devis_id', $request->devis_id)->where('produit_id', $request->produit)->first(['id'])->id)) {
+                        return back()->with('danger', "Produit deja enregistrer");
+                    } else {
+                        $data = [
+                            'devis_id'              => $request->devis_id,
+                            'produit_id'               => $request->produit,
+                            'quantite_demandee'       => $request->qte,
+                            'prix_unitaire_demande'  => $request->prix,
+                        ];
+                        DetailDevis::create($data);
 
-                            $stock->update([
-                                'quantite_en_stock'=>$stock->quantite_en_stock-$request->qte,
-                            ]);
+                        $stock->update([
+                            'quantite_en_stock' => $stock->quantite_en_stock - $request->qte,
+                        ]);
 
-                            return back();
-                        }
+                        return back();
+                    }
                 }
-                return back()->with('danger',"La quantité stock est insuffisante");
+                return back()->with('danger', "La quantité stock est insuffisante");
             }
-            return back()->with('danger',"Vous n'avez pas de produit dans l'entrepot");;
+            return back()->with('danger', "Vous n'avez pas de produit dans l'entrepot");;
 
 
 
 
 
-                // $devis=devis::find($request->devis_id);
+            // $devis=devis::find($request->devis_id);
 
-                // $devis->update([
-                //     'fournisseur_id' =>$request->fournisseur,
-                //     'montant_total' =>$request->montant_ht,
-                //     'etat' =>'en cours',
-                // ]);
-                // return back();
-                // return redirect('detail_devis/'.$devis->id.'/show');
+            // $devis->update([
+            //     'fournisseur_id' =>$request->fournisseur,
+            //     'montant_total' =>$request->montant_ht,
+            //     'etat' =>'en cours',
+            // ]);
+            // return back();
+            // return redirect('detail_devis/'.$devis->id.'/show');
 
-        }
-            return redirect('/auth')->with('danger',"Session expirée");
 
-            //
-            // if(Auth::check()){
 
             //
             // dd($request->devis_id,$request->produit,
@@ -133,9 +125,8 @@ class DetailDevisController extends Controller
             //     ]);
             //     return redirect('detail_devis/'.$devis->id.'/show');
 
-            // }
-            // return redirect('/auth')->with('danger',"Session expirée");
-
+        }
+        return redirect('/')->with('danger', "Session expirée");
     }
 
     /**
@@ -143,14 +134,17 @@ class DetailDevisController extends Controller
      */
     public function show(string $id)
     {
-        //
-        $id=decrypt($id);
-         $devis=Devis::find($id);
-         $agence_id=Auth::user()->agence_id;
-         $detail_deviss=DetailDevis::where('devis_id',$devis->id)->get();
-         $activite_investissements=ActiviteInvestissement::where("agence_id",$agence_id)->where('etat_activite','valider')->get();
-         $total_ht=DetailDevis::where('devis_id',$devis->id)->selectRaw('sum(quantite_demandee*prix_unitaire_demande) as total')->first('total');
-         return view('e-commerce.devis_encours', compact('devis','detail_deviss','total_ht','activite_investissements'));
+        if (Auth::check()) {
+            $id = decrypt($id);
+            $devis = Devis::find($id);
+            $agence_id = Auth::user()->agence_id;
+            $detail_deviss = DetailDevis::where('devis_id', $devis->id)->get();
+            $activite_investissements = ActiviteInvestissement::where("agence_id", $agence_id)->where('etat_activite', 'valider')->get();
+            $activite_investissement = ActiviteInvestissement::find($devis->activite_id);
+            $total_ht = DetailDevis::where('devis_id', $devis->id)->selectRaw('sum(quantite_demandee*prix_unitaire_demande) as total')->first('total');
+            return view('e-commerce.devis_encours', compact('devis', 'detail_deviss', 'total_ht', 'activite_investissements', 'activite_investissement'));
+        }
+        return redirect('/')->with('danger', "Session expirée");
     }
 
     /**
@@ -158,15 +152,17 @@ class DetailDevisController extends Controller
      */
     public function edit(string $id)
     {
-        $id=decrypt($id);
-        $devis=Devis::find($id);
-        $agence_id=Auth::user()->agence_id;
-        $produits=Produit::where('agence_id',$agence_id)->get();
-        $clients=Client::all();
-        $detail_deviss=DetailDevis::where('devis_id',$devis->id)->get();
-        $total_ht=DetailDevis::where('devis_id',$devis->id)->selectRaw('sum(quantite_demandee*prix_unitaire_demande) as total')->first('total');
-        return view('e-commerce.edit_devis', compact('produits','clients','devis','detail_deviss','total_ht'));
-
+        if (Auth::check()) {
+            $id = decrypt($id);
+            $devis = Devis::find($id);
+            $agence_id = Auth::user()->agence_id;
+            $produits = Produit::where('agence_id', $agence_id)->get();
+            $clients = Client::all();
+            $detail_deviss = DetailDevis::where('devis_id', $devis->id)->get();
+            $total_ht = DetailDevis::where('devis_id', $devis->id)->selectRaw('sum(quantite_demandee*prix_unitaire_demande) as total')->first('total');
+            return view('e-commerce.edit_devis', compact('produits', 'clients', 'devis', 'detail_deviss', 'total_ht'));
+        }
+        return redirect('/')->with('danger', "Session expirée");
     }
 
     /**
@@ -182,50 +178,53 @@ class DetailDevisController extends Controller
      */
     public function destroy(string $id)
     {
-        $id=decrypt($id);
+        if (Auth::check()) {
+            $id = decrypt($id);
 
-        $produit=DetailDevis::find($id);
-        $stock=StockProduitActivite::where('produit_id',$produit->produit_id)->first();
+            $produit = DetailDevis::find($id);
+            $stock = StockProduitActivite::where('produit_id', $produit->produit_id)->first();
 
-        // dd($produit, $stock);
+            // dd($produit, $stock);
 
-        $stock->update([
-            'quantite_en_stock'=>$stock->quantite_en_stock+$produit->quantite_demandee,
-        ]);
+            $stock->update([
+                'quantite_en_stock' => $stock->quantite_en_stock + $produit->quantite_demandee,
+            ]);
 
-        $produit->delete();
+            $produit->delete();
 
-        return back();
+            return back();
+        }
+        return redirect('/')->with('danger', "Session expirée");
     }
 
     public function client_devis(Request $request)
     {
+        if (Auth::check()) {
+            $societe_id = Auth::user()->societe_id;
+            $tel = $request->id;
 
-        $societe_id=Auth::user()->societe_id;
-        $tel=$request->id;
+            if (isset(Client::where('telephone', $tel)->where('societe_id', $societe_id)->first(['id'])->id)) {
 
-        if(isset(Client::where('telephone' ,$tel)->where('societe_id',$societe_id)->first(['id'])->id)){
+                $agence_id = Auth::user()->agence_id;
 
-            $agence_id=Auth::user()->agence_id;
+                $id = Client::where('telephone', $tel)->where('societe_id', $societe_id)->first(['id'])->id;
 
-            $id=Client::where('telephone' ,$tel)->where('societe_id',$societe_id)->first(['id'])->id;
+                $data['client'] = Client::where('id', $id)->get();
+                return response()->json($data);
+            } else {
 
-            $data['client']=Client::where('id',$id)->get();
-            return response()->json($data);
-
-        }else{
-
-            Client::create([
-                'telephone'=>$tel,
-                'societe_id'=>$societe_id,
-            ]);
-            /**
-             * si le telephone existe afficher le client
-             */
-            $id=Client::where('telephone' ,$tel)->where('societe_id',$societe_id)->first(['id'])->id;
-            $data['client']=Client::where('id',$id)->get();
-            return response()->json($data);
-
+                Client::create([
+                    'telephone' => $tel,
+                    'societe_id' => $societe_id,
+                ]);
+                /**
+                 * si le telephone existe afficher le client
+                 */
+                $id = Client::where('telephone', $tel)->where('societe_id', $societe_id)->first(['id'])->id;
+                $data['client'] = Client::where('id', $id)->get();
+                return response()->json($data);
+            }
         }
+        return redirect('/')->with('danger', "Session expirée");
     }
 }
